Disable distribution if we can't find coef node
If the expression being distributed is of form:
a += -b * tmp_xx
The uminus node is as of now not stringified and hence the coef
node in this case (-b) is not found and we end up trying to create
binary node with one of the operand as nullptr.
Added assert so that in future binary node constructor will always
make sure both operands are valid.